speed startup: alter recovery loop
We do a recovery on startup.  But the code does:
   Sleep for ctdb->tunable.recover_interval.
   Check for recovery.
   
We want to do it in the other order.  This is best done by extracting
the loop into a separate "main_loop" function.

Seconds between ctdbd first log message and node healthy:
BEFORE:	24.09
AFTER: 23.58

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>


(This used to be ctdb commit 097046025176b9fcb670839d1a9f100f890e7ed2)