speed startup: don't wait a full recovery interval if we've already waited
We currently sleep for one second, whether or not we've already slept.
Change this to sleep for the remainder of the second, if at all.

Seconds between ctdbd first log message and node healthy:
BEFORE:	18.09
AFTER: 17.08

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>


(This used to be ctdb commit fde760b5f39c77172308a583da4c2443b71541c9)