s3:smbd: add a nfs backend for sysquotas.
This module is based on the Solaris/FreeBSD implementation
of NFS quotas in the quotas.c module.

It implements the SMB_USER_QUOTA_TYPE query of the
get_quotas call. The other types and the set_quota call
are not implemented.
(cherry picked from commit 3add3aa17de7954b01e2c9273295195438ac30e5)