Factor out the recent changes into a function - check_parent_exists(). Fix this to ensure that if "start" is manipulated, then "dirpath" is changed also.
Ensures that when the path:

/a/long/file/name/path.txt

is processed, we first stat:

/a/long/file/name/path.txt

and if this fails, we try to stat:

/a/long/file/name

if this path exists (the normal case when creating a new
entry in a directory) then we no longer do the individual
path name walk, but only do case insensitive lookup on the
last component. If the stat fails we do the full pathname
walk as normal in 3.5.x and below. Metze, examine this
change for your back-port.

Jeremy.
(cherry picked from commit 1787c1dfc3a506c0afee6d2bb67ba8789c709b86)