s4:irpc: add padding to the IPRC header for 8 byte alignment
As we marshall the iprc header and the payload to the ndr_push
context, we should pad the irpc header to let the payload start
at an 8 byte boundary.

This way we get the alignment still be correct if we remove the
header before passing the raw payload to the caller.

As we use IRPC more and more for complex NDR marshalled structures
we need to get this right, in order to not get random ndr_pull failures.

metze

Signed-off-by: Anatoliy Atanasov <anatoliy.atanasov@postpath.com>