s4-heimdal_build fix up build after heimdal import
Heimdal has it's own dynconfig like system, and so we need the
LIBDIR etc in the CFLAGS here.

We also need to define build rules for the new files imported in
the merge commit, and fix up some header files.

This includes the work by Matthieu Patou <mat@matws.net>

Andrew Bartlett